Seperated out user and session stores
